+++ CHANGES.rst
@@ -28,6 +28,7 @@ Changelog
 
 -   Move RAVEN tasks to main worker and make GCN ingestion sequential.
 
+-   Limit the number of MDC BBH bilby jobs to less than 5 to save disk usage.
 
 2.1.1 "Lone Island Mountain Devil" (2023-05-29)
 -----------------------------------------------